fix: avoid trying to copy shapes that have been moved by the word-to-dtbook script

+++ Common/DaisyConverterLib/Pipeline/Pipeline2/Scripts/WordToDtbook.cs
@@ -293,7 +293,15 @@ namespace Daisy.SaveAsDAISY.Conversion.Pipeline.Pipeline2.Scripts
             if(ExtractedShapes.Count > 0) {
                 foreach (string shape in ExtractedShapes)
                 {
-                    File.Copy(shape, Path.Combine(Parameters["output"].Value.ToString(), Path.GetFileName(shape)), true);
+                    if(File.Exists(shape) == false) {
+                        // File has been already copied by the new word to dtbook script, continue parsing
+                        continue;
+                    }
+                    try {
+                        File.Copy(shape, Path.Combine(Parameters["output"].Value.ToString(), Path.GetFileName(shape)), true);
+                    } catch(Exception ex) {
+                        EventsHandler.onProgressMessageReceived(this, new DaisyEventArgs("Error while copying extracted shape "+shape+": "+ex.Message));
+                    }
                 }
             }
         }
